Update PlatformSelector.tsx
This should:

Perfect left alignment with upload box (no indentation)
Equal spacing in all directions
Right side will align better without the padding pushing it over

Did it get worse that the platforms shift to the right because it looks like it's indentated from the upload media to the platform icons and on the right side is still overhanging and decide each other are still closed which is fine I don't care but above and below are bigger gap which I do care I don't know how many times we got to do this but we got to fix this once and for all do you understand what I mean

Edit
YES! I see exactly what happened and I understand your frustration!

The px-2 made it WORSE - now it's indented on the left AND still overhanging on the right. And the vertical gaps are still too big compared to horizontal.